+++ src/commands/identify-characters.ts
@@ -13,15 +13,7 @@ export const identifyCharacters: CommandCallback = async (editor) =>
 	if (text === undefined)
 		return;
 
-	const message = Array.from(text)
-		.map(char => ({
-			char,
-			code: char.codePointAt(0) ? char.codePointAt(0)! : char.charCodeAt(0)
-		}))
-		.map(item => ({
-			...item,
-			entry: data.find(d => d.codes.length === 1 && d.codes[0] === item.code)
-		}))
+	const message = identify(text)
 		.map(item => item.entry === undefined
 			? `Unknown character code: ${codesToHex([item.code])}`
 			: `${item.char}: ${item.entry.name} (${codesToHex([item.code])})`)
@@ -39,3 +31,24 @@ export const identifyCharacters: CommandCallback = async (editor) =>
 		await window.showTextDocument(document, ViewColumn.Two);
 	}
 };
+
+/**
+ * Identifies Unicode characters of a string.
+ * @param text The text whose characters should be identified.
+ * @returns
+ *     List of characters and their Unicode entries.
+ *     `entry` is `undefined` if it could not be found.
+ */
+export function identify(text: string)
+{
+	return Array.from(text)
+		.map(char => ({
+			char,
+			code: char.codePointAt(0) ? char.codePointAt(0)! : char.charCodeAt(0)
+		}))
+		.map(item => ({
+			...item,
+			entry: data.find(d => d.codes.length === 1 && d.codes[0] === item.code)
+		}));
+}
+